requirements.txt
customtkinter==5.2.2
mysql-connector-python==9.1.0
bcrypt==4.2.0
